requirements.txt
blinker==1.8.2
click==8.1.7
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4
numpy==2.1.2
pillow==10.4.0
setuptools==75.1.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
Werkzeug==3.0.4
